Make the preferred radius configurable on the Settings screen

+++ app/src/main/java/com/tarasfedyk/lunchplaces/store/GeoRepoImpl.kt
@@ -53,7 +53,7 @@ class GeoRepoImpl @Inject constructor(
             RankingCriterion.DISTANCE -> SearchByTextRequest.RankPreference.DISTANCE
         }
         val circularBounds = CircularBounds.newInstance(
-            searchFilter.originPoint, searchFilter.settings.preferredRadius
+            searchFilter.originPoint, searchFilter.settings.preferredRadius.toDouble()
         )
 
         val cancellationTokenSource = CancellationTokenSource()
